Refactoring project

+++ src/main/java/com/gp/solutions/controller/PartyController.java
@@ -1,9 +1,14 @@
 package com.gp.solutions.controller;
 
+import java.security.Principal;
 import java.util.Collection;
 
 import com.gp.solutions.entity.dbo.Party;
+import com.gp.solutions.entity.dbo.User;
 import com.gp.solutions.repository.PartyRepository;
+import com.gp.solutions.repository.UserRepository;
+import com.gp.solutions.type.UserRole;
+import io.swagger.annotations.Api;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,38 +20,52 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping(PartyController.REQUEST_MAPPING)
+@Api(description = "CRUD operations for party")
 public class PartyController {
 
-	public static final String REQUEST_MAPPING = "/parties";
+    public static final String REQUEST_MAPPING = "/parties";
 
-	@Autowired
-	private PartyRepository partyRepo;
+    @Autowired
+    private PartyRepository partyRepo;
 
-	@RequestMapping(method = RequestMethod.GET)
-	public ResponseEntity<Collection<Party>> getParties() {
-		return new ResponseEntity<>(partyRepo.findAll(), HttpStatus.OK);
-	}
+    @Autowired
+    private UserRepository userRepository;
 
-	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public ResponseEntity<Party> getParty(@PathVariable long id) {
-		Party party = partyRepo.findOne(id);
+    @RequestMapping(method = RequestMethod.GET)
+    public ResponseEntity<Collection<Party>> getParties() {
+        return new ResponseEntity<>(partyRepo.findAll(), HttpStatus.OK);
+    }
 
-		if (party != null) {
-			return new ResponseEntity<>(partyRepo.findOne(id), HttpStatus.OK);
-		} else {
-			return new ResponseEntity<>((Party) null, HttpStatus.NOT_FOUND);
-		}
-	}
+    @RequestMapping(value = "/{id}", method = RequestMethod.GET)
+    public ResponseEntity<Party> getParty(@PathVariable long id) {
+        final Party party = partyRepo.findOne(id);
 
-	@RequestMapping(method = RequestMethod.POST)
-	public ResponseEntity<Party> addParty(@RequestBody Party party) {
-		return new ResponseEntity<>(partyRepo.save(party), HttpStatus.CREATED);
-	}
+        if (party != null) {
+            return new ResponseEntity<>(partyRepo.findOne(id), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>((Party) null, HttpStatus.NOT_FOUND);
+        }
+    }
 
-	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<Void> deleteParty(@PathVariable long id) {
-		partyRepo.delete(id);
-		return new ResponseEntity<>(HttpStatus.OK);
-	}
+    @RequestMapping(method = RequestMethod.POST)
+    public ResponseEntity<Party> addParty(@RequestBody Party party, Principal principal) {
+        final User currentUser = userRepository.findOneByUsername(principal.getName());
+        if (UserRole.ADMIN.equals(currentUser.getUserRole())) {
+            return new ResponseEntity<>(partyRepo.save(party), HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+    }
+
+    @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
+    public ResponseEntity<Void> deleteParty(@PathVariable long id, Principal principal) {
+        final User currentUser = userRepository.findOneByUsername(principal.getName());
+        if (UserRole.ADMIN.equals(currentUser.getUserRole())) {
+            partyRepo.delete(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+    }
 
 }
